Move FormatOptions to vector module
