+++ crates/harp/src/vector/numeric_vector.rs
@@ -14,7 +14,7 @@ use libr::REAL_ELT;
 use libr::SEXP;
 
 use crate::object::RObject;
-use crate::vector::formatted_vector::FormatOptions;
+use crate::vector::FormatOptions;
 use crate::vector::Vector;
 
 #[harp_macros::vector]
